requirements.txt
torch>=2.4.0
wandb
scipy
pandas
xgboost
scikit-learn
numpy>=1.24,<3.0
matplotlib
ipykernel
pyyaml

